Add Color Picker Client App, integrate SimpleColorPicker.
Add a Color Picker client app with a BColorWell control and also integrate
the simpleColorPicker into it. This involves lots of changes and breaks the
other color pickers for now, but, also makes the simple color picker work in
a client server relationship that will ultimately mean a more modular color
picker selection once the other color pickers are updated.